Add admission and discharge status to the history. #497 (Software engineering assignment) Group 3
(bug)The problem is the absence of admission and discharge status in the history, which hinders tracking a patient's care timeline and continuity.
(Solution) We added on  ngOnInit() function a parameter <this,params> which takes data from this.visits$<this.params = {
      visits: [
        { visit: this.visits$ },
      ],
    };>
We then added function patientStatus() which uses this,params to access the visit.visit.stopDatetime and visit.visit.startDatetime, So when visit.visit.stopDatetime isnt available it displays STATUS of admitted when both visit.visit.stopDatetime && visit.visit.startDatetime it displays status of Discharged. then added div that displays result
2.then added in <patient-history.component.html> a span in which changes admitted and discharge

GROUP MEMBERS
1 James c Sawe 2022-04-12029
2-HAPPINESS NUIS KASAMBALA 2022-04-03745
3-Sonia Moses limbu 2022-04-05039
4-2022-04-01303 Caren A Christopher
5-2022-04-06412 Karenbhoke C Maseke

Co-Authored-By: HAPPYNUIS <167102138+happynuis@users.noreply.github.com>
Co-Authored-By: Karenbhoke <167118734+karenbhoke@users.noreply.github.com>
Co-Authored-By: carenchris <139714150+carenchris@users.noreply.github.com>
Co-Authored-By: James Cuthbert Sawe <jamescuthbert907@gmail.com>