Fixing 255 bug
The Problem:

A table of issuing items under main store was displaying fewer items than the number selected by the user. For example, when the user selected 10 items per page, the table displayed fewer than 10 rows. Upon investigation, we discovered that the issue was caused by an *ngIf directive in the HTML template. This directive was preventing the rendering of items with a status of "DRAFT".

How the *ngIf Worked:

The *ngIf directive in Angular conditionally includes or excludes elements from the DOM based on a boolean expression. In this case, the directive was likely written as follows:

<mat-expansion-panel
          class="border mat-elevation-z0 mt-2"
         *ngIf="
            issue?.requisitionStatuses[issue?.requisitionStatuses?.length - 1]
              ?.status !== 'DRAFT'
          "
          [expanded]="
            viewIssueItems && viewIssueItems === issue?.uuid
              ? true
              : issueIndex === 0
              ? true
              : false
          "
        >
           ...
        </mat-expansion-panel>

This condition excluded items with a status of "DRAFT" from being rendered in the table. As a result, even though the total data set contained the desired number of items, the presence of "DRAFT" items caused the table to display fewer rows than expected.

The Solution:

To resolve the issue, the *ngIf directive that filtered out items with a "DRAFT" status was removed. The table was modified as follows:

<mat-expansion-panel
          class="border mat-elevation-z0 mt-2"
          [expanded]="
            viewIssueItems && viewIssueItems === issue?.uuid
              ? true
              : issueIndex === 0
              ? true
              : false
          "
        >
           ...
        </mat-expansion-panel>

By removing the condition, all items, regardless of their status, were rendered in the table. This allowed the table to consistently display the exact number of items selected by the user.

Participants Registration Numbers:
2022-04-11894
2022-04-04624
2022-04-07307
2022-04-04998
2022-04-09788